Add files via upload
“Development of Application of Image Detection using Python”, presents a robust image detection system designed to identify and track vehicles using deep learning techniques.

The core objective of the project is to build an application that leverages the YOLOv8 (You Only Look Once) object detection model to detect, track, and analyze moving vehicles in traffic surveillance footage. Developed in Python, the system integrates powerful libraries such as OpenCV, Tkinter, and Pandas to create a complete GUI-based solution capable of real-time detection, vehicle counting, speed estimation, and data logging.

The application defines a region of interest (ROI) to isolate the traffic zone and utilizes time-based logic to calculate the speed of each vehicle as it crosses virtual detection lines. Each detected vehicle is assigned a unique ID for accurate tracking and data collection. The final report of vehicle movement, including IDs, crossing times, and speed, is generated and can be exported to Excel for further analysis.

This system provides a practical and scalable solution for intelligent traffic monitoring and road safety enforcement, with potential applications in smart city infrastructure, highway surveillance, and automated traffic control systems.